Extract common server stuff
