+++ monadoc.cabal
@@ -79,6 +79,7 @@ library
     Monadoc.Ghc
     Monadoc.Main
     Monadoc.Server.Application
+    Monadoc.Server.Common
     Monadoc.Server.Main
     Monadoc.Server.Middleware
     Monadoc.Server.Settings
@@ -125,6 +126,7 @@ test-suite test
     Monadoc.GhcSpec
     Monadoc.MainSpec
     Monadoc.Server.ApplicationSpec
+    Monadoc.Server.CommonSpec
     Monadoc.Server.MainSpec
     Monadoc.Server.MiddlewareSpec
     Monadoc.Server.SettingsSpec
