+++ src/lib/Monadoc/Server/Middleware.hs
@@ -9,6 +9,7 @@ import qualified Data.Map as Map
 import qualified Data.Maybe as Maybe
 import qualified GHC.Clock as Clock
 import qualified Monadoc.Console as Console
+import qualified Monadoc.Server.Common as Common
 import qualified Monadoc.Server.Settings as Settings
 import qualified Monadoc.Type.Config as Config
 import qualified Monadoc.Utility.Utf8 as Utf8
@@ -53,7 +54,7 @@ handleEtag handle request respond = handle request $ \response ->
     hasEtag = Maybe.isJust expected
     actual = lookup Http.hETag $ Wai.responseHeaders response
   in respond $ if isGet && isSuccessful && hasEtag && actual == expected
-    then Settings.responseBS
+    then Common.responseBS
       Http.notModified304
       (Map.fromList $ Wai.responseHeaders response)
       ByteString.empty
