+++ src/lib/Monadoc/Server/Settings.hs
@@ -1,29 +1,20 @@
 module Monadoc.Server.Settings
-  ( Headers
-  , defaultHeaders
-  , fromConfig
+  ( fromConfig
   , onException
   , onExceptionResponse
-  , responseBS
-  , statusResponse
-  , stringResponse
   )
 where
 
 import qualified Control.Monad as Monad
 import qualified Control.Monad.Catch as Exception
-import qualified Crypto.Hash as Crypto
 import qualified Data.ByteString as ByteString
-import qualified Data.ByteString.Lazy as LazyByteString
-import qualified Data.List as List
-import qualified Data.Map as Map
 import qualified Monadoc.Console as Console
 import qualified Monadoc.Data.Commit as Commit
 import qualified Monadoc.Data.Version as Version
+import qualified Monadoc.Server.Common as Common
 import qualified Monadoc.Type.Config as Config
 import qualified Monadoc.Utility.Utf8 as Utf8
 import qualified Network.HTTP.Types as Http
-import qualified Network.HTTP.Types.Header as Http
 import qualified Network.Wai as Wai
 import qualified Network.Wai.Handler.Warp as Warp
 
@@ -54,48 +45,8 @@ onException _ someException@(Exception.SomeException exception) =
 
 onExceptionResponse :: Config.Config -> Exception.SomeException -> Wai.Response
 onExceptionResponse config _ =
-  statusResponse Http.internalServerError500 $ defaultHeaders config
-
-type Headers = Map.Map Http.HeaderName ByteString.ByteString
-
-statusResponse :: Http.Status -> Headers -> Wai.Response
-statusResponse status headers = stringResponse status headers $ unwords
-  [show $ Http.statusCode status, Utf8.toString $ Http.statusMessage status]
-
-stringResponse :: Http.Status -> Headers -> String -> Wai.Response
-stringResponse status headers =
-  responseBS
-      status
-      (Map.insert Http.hContentType "text/plain;charset=utf-8" headers)
-    . Utf8.fromString
-
-responseBS :: Http.Status -> Headers -> ByteString.ByteString -> Wai.Response
-responseBS status headers body =
-  let
-    contentLength = Utf8.fromString . show $ ByteString.length body
-    etag = Utf8.fromString . show . show $ Crypto.hashWith Crypto.SHA256 body
-    extraHeaders =
-      Map.fromList [(Http.hContentLength, contentLength), (Http.hETag, etag)]
-  in Wai.responseLBS
-    status
-    (Map.toList $ Map.union extraHeaders headers)
-    (LazyByteString.fromStrict body)
-
-defaultHeaders :: Config.Config -> Headers
-defaultHeaders config = Map.fromList
-  [ ("Content-Security-Policy", contentSecurityPolicy)
-  , ("Referrer-Policy", "no-referrer")
-  , ("Strict-Transport-Security", strictTransportSecurity config)
-  , ("X-Content-Type-Options", "nosniff")
-  , ("X-Frame-Options", "deny")
-  ]
-
-contentSecurityPolicy :: ByteString.ByteString
-contentSecurityPolicy = "base-uri 'none', default-src 'self'"
-
-strictTransportSecurity :: Config.Config -> ByteString.ByteString
-strictTransportSecurity config =
-  if List.isPrefixOf "https:" $ Config.url config then "86400" else "0"
+  Common.statusResponse Http.internalServerError500
+    $ Common.defaultHeaders config
 
 serverName :: ByteString.ByteString
 serverName =
