+++ src/test/Monadoc/Server/CommonSpec.hs
@@ -1,0 +1,11 @@
+module Monadoc.Server.CommonSpec
+  ( spec
+  )
+where
+
+import qualified Test.Hspec as Hspec
+
+spec :: Hspec.Spec
+spec = Hspec.describe "Monadoc.Server.Common" $ do
+
+  Hspec.it "needs tests" Hspec.pending
